@@ -308,17 +278,18 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>2.2</version>
+                <version>3.7.0</version>
                 <configuration>
-                    <source>1.6</source>
-                    <target>1.6</target>
+                    <source>1.8</source>
+                    <target>1.8</target>
                 </configuration>
             </plugin>
 	    <plugin>
 	      <groupId>org.apache.maven.plugins</groupId>
 	      <artifactId>maven-javadoc-plugin</artifactId>
+	      <version>3.0.0</version>
 	      <configuration>
-		<additionalparam>-Xdoclint:none</additionalparam>
+		      <additionalparam>-Xdoclint:none</additionalparam>
 	      </configuration>
 	    </plugin>
 
