@@ -24,10 +24,14 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.GregorianCalendar;
 import java.util.List;
+import java.util.Objects;
 
 import javax.xml.datatype.DatatypeConfigurationException;
 import javax.xml.datatype.DatatypeFactory;
 import javax.xml.datatype.XMLGregorianCalendar;
+
+import org.junit.Assert;
+import org.junit.Test;
 
 import ma.glasnost.orika.MapperFacade;
 import ma.glasnost.orika.MapperFactory;
@@ -35,12 +39,6 @@ import ma.glasnost.orika.converter.builtin.CloneableConverter;
 import ma.glasnost.orika.converter.builtin.PassThroughConverter;
 import ma.glasnost.orika.impl.DefaultMapperFactory;
 import ma.glasnost.orika.test.MappingUtil;
-
-import org.apache.commons.lang.builder.EqualsBuilder;
-import org.apache.commons.lang.builder.HashCodeBuilder;
-import org.junit.Test;
-
-import org.junit.Assert;
 
 public class CloneableConverterTestCase {
     
@@ -187,11 +185,11 @@ public class CloneableConverterTestCase {
         }
         
         public boolean equals(Object that) {
-        	return EqualsBuilder.reflectionEquals(this, that);
+        	return Objects.equals(id, ((SampleCloneable) that).id);
         }
         
         public int hashCode() {
-        	return HashCodeBuilder.reflectionHashCode(this);
+        	return Objects.hash(id);
         }
         
     }
