@@ -135,7 +135,7 @@ public class BeanToMultiOccurrenceTestCase {
                     person.birthDate.toXMLFormat()));
         }
         
-        MapOfPeople mapBack = mapper.map(result, MapOfPeople.class);
+        mapper.map(result, MapOfPeople.class);
     }
     
     /**
