@@ -21,6 +21,7 @@ package ma.glasnost.orika.test.generics;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
 
 import ma.glasnost.orika.MapperFactory;
 import ma.glasnost.orika.converter.builtin.PassThroughConverter;
@@ -30,8 +31,6 @@ import ma.glasnost.orika.metadata.ClassMapBuilder;
 import ma.glasnost.orika.metadata.Type;
 import ma.glasnost.orika.metadata.TypeBuilder;
 
-import org.apache.commons.lang.builder.EqualsBuilder;
-import org.apache.commons.lang.builder.HashCodeBuilder;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -50,11 +49,11 @@ public class GenericCollectionsTestCase {
         }
         
         public boolean equals(Object that) {
-            return EqualsBuilder.reflectionEquals(this, that);
+        	return Objects.equals(name, ((Person) that).name);
         }
         
         public int hashCode() {
-            return HashCodeBuilder.reflectionHashCode(this);
+        	return Objects.hash(name);
         }
     }
     
