@@ -22,12 +22,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.junit.Assert;
+import org.junit.Test;
 
 import ma.glasnost.orika.impl.UtilityResolver;
 import ma.glasnost.orika.metadata.Property;
-import ma.glasnost.orika.property.IntrospectorPropertyResolver;
-
-import org.junit.Test;
 
 public class VariableRefTestCase {
     
